requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
httpx>=0.24.0,<0.25.0
aiohttp>=3.8.0,<3.9.0
redis>=4.5.0,<4.6.0
python-dotenv>=0.19.0,<0.20.0
pydantic>=1.8.0,<2.0.0
prometheus-client>=0.12.0,<0.13.0
pytest>=7.0.0,<7.1.0
pytest-asyncio>=0.18.0,<0.19.0
numpy>=1.21.0,<1.22.0
pandas>=1.3.0,<1.4.0  # For future ML analysis
scikit-learn>=0.24.0,<0.25.0  # For future ML components 